[PhotosUI] Move macOS photosui bindings out of photos.cs into photosui.cs
