+++ src/frameworks.sources
@@ -1606,6 +1606,7 @@ MAC_FRAMEWORKS =            \
 	OpenGL                  \
 	PdfKit                  \
 	Photos                  \
+	PhotosUI				\
 	PrintCore		\
 	QTKit                   \
 	QuartzComposer          \
