+++ src/photosui.cs
@@ -1,12 +1,16 @@
 using XamCore.CoreGraphics;
 using XamCore.ObjCRuntime;
 using XamCore.Foundation;
+#if !MONOMAC
 using XamCore.UIKit;
+#else
+using XamCore.AppKit;
+#endif
 using XamCore.Photos;
 using System;
 
-namespace XamCore.PhotosUI
-{
+namespace XamCore.PhotosUI {
+#if !MONOMAC
 	[NoTV]
 	[iOS (8,0)]
 	[Protocol]
@@ -87,4 +91,232 @@ namespace XamCore.PhotosUI
 		[Export ("livePhotoView:didEndPlaybackWithStyle:")]
 		void DidEndPlayback (PHLivePhotoView livePhotoView, PHLivePhotoViewPlaybackStyle playbackStyle);
 	}
+#else
+	[Mac (10,13, onlyOn64: true)][NoiOS][NoTV][NoWatch]
+	[Static]
+	interface PHProjectType {
+		[Field ("PHProjectTypeUndefined")]
+		NSString Undefined { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[DisableDefaultCtor]
+	[BaseType (typeof (NSExtensionContext))]
+	interface PHProjectExtensionContext : NSSecureCoding, NSCopying {
+
+		[Export ("photoLibrary")]
+		PHPhotoLibrary PhotoLibrary { get; }
+
+		[Export ("project")]
+		PHProject Project { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[DisableDefaultCtor]
+	[BaseType (typeof (PHProjectElement))]
+	interface PHProjectJournalEntryElement : NSSecureCoding {
+
+		[Export ("date")]
+		NSDate Date { get; }
+
+		[NullAllowed, Export ("assetElement")]
+		PHProjectAssetElement AssetElement { get; }
+
+		[NullAllowed, Export ("textElement")]
+		PHProjectTextElement TextElement { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[DisableDefaultCtor]
+	[BaseType (typeof (PHProjectElement))]
+	interface PHProjectTextElement : NSSecureCoding {
+
+		[Export ("text")]
+		string Text { get; }
+
+		[NullAllowed, Export ("attributedText")]
+		NSAttributedString AttributedText { get; }
+
+		[Export ("textElementType")]
+		PHProjectTextElementType TextElementType { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[Protocol]
+	interface PHProjectExtensionController {
+
+		[Export ("supportedProjectTypes", ArgumentSemantic.Copy)]
+		PHProjectTypeDescription[] SupportedProjectTypes { get; }
+
+		[Abstract]
+		[Export ("beginProjectWithExtensionContext:projectInfo:completion:")]
+		void BeginProject (PHProjectExtensionContext extensionContext, PHProjectInfo projectInfo, Action<NSError> completion);
+
+		[Abstract]
+		[Export ("resumeProjectWithExtensionContext:completion:")]
+		void ResumeProject (PHProjectExtensionContext extensionContext, Action<NSError> completion);
+
+		[Abstract]
+		[Export ("finishProjectWithCompletionHandler:")]
+		void FinishProject (Action completion);
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[BaseType (typeof (NSObject))]
+	[DisableDefaultCtor]
+	interface PHProjectTypeDescription : NSSecureCoding {
+
+		[Export ("projectType")]
+		NSString ProjectType { get; }
+
+		[Export ("localizedTitle")]
+		string LocalizedTitle { get; }
+
+		[NullAllowed, Export ("localizedDescription")]
+		string LocalizedDescription { get; }
+
+		[NullAllowed, Export ("image", ArgumentSemantic.Copy)]
+		NSImage Image { get; }
+
+		[Export ("subtypeDescriptions", ArgumentSemantic.Copy)]
+		PHProjectTypeDescription[] SubtypeDescriptions { get; }
+
+		[Export ("initWithProjectType:title:description:image:subtypeDescriptions:")]
+		[DesignatedInitializer]
+		IntPtr Constructor (NSString projectType, string localizedTitle, [NullAllowed] string localizedDescription, [NullAllowed] NSImage image, PHProjectTypeDescription[] subtypeDescriptions);
+
+		[Export ("initWithProjectType:title:description:image:")]
+		IntPtr Constructor (NSString projectType, string localizedTitle, [NullAllowed] string localizedDescription, [NullAllowed] NSImage image);
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[BaseType (typeof (NSObject))]
+	[DisableDefaultCtor]
+	interface PHProjectRegionOfInterest : NSSecureCoding {
+
+		[Export ("rect")]
+		CGRect Rect { get; }
+
+		[Export ("weight")]
+		double Weight { get; }
+
+		[Export ("identifier")]
+		string Identifier { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[BaseType (typeof (NSObject))]
+	[DisableDefaultCtor]
+	interface PHProjectElement : NSSecureCoding {
+
+		[Export ("weight")]
+		double Weight { get; }
+
+		[Export ("placement")]
+		CGRect Placement { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[DisableDefaultCtor]
+	[BaseType (typeof (PHProjectElement))]
+	interface PHProjectAssetElement : NSSecureCoding {
+
+		[Export ("cloudAssetIdentifier")]
+		PHCloudIdentifier CloudAssetIdentifier { get; }
+
+		[Export ("annotation")]
+		string Annotation { get; }
+
+		[Export ("cropRect")]
+		CGRect CropRect { get; }
+
+		[Export ("regionsOfInterest")]
+		PHProjectRegionOfInterest[] RegionsOfInterest { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[BaseType (typeof (NSObject))]
+	[DisableDefaultCtor]
+	interface PHProjectInfo : NSSecureCoding {
+
+		[Export ("creationSource")]
+		PHProjectCreationSource CreationSource { get; }
+
+		[Export ("projectType")]
+		NSString ProjectType { get; }
+
+		[Export ("sections")]
+		PHProjectSection[] Sections { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[BaseType (typeof (NSObject))]
+	[DisableDefaultCtor]
+	interface PHProjectSection : NSSecureCoding {
+
+		[Export ("sectionContents")]
+		PHProjectSectionContent[] SectionContents { get; }
+
+		[Export ("sectionType")]
+		PHProjectSectionType SectionType { get; }
+
+		[Export ("title")]
+		string Title { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[BaseType (typeof (NSObject))]
+	[DisableDefaultCtor]
+	interface PHProjectSectionContent : NSSecureCoding {
+
+		[Export ("elements")]
+		PHProjectElement[] Elements { get; }
+
+		[Export ("numberOfColumns")]
+		nint NumberOfColumns { get; }
+
+		[Export ("aspectRatio")]
+		double AspectRatio { get; }
+
+		[Export ("cloudAssetIdentifiers")]
+		PHCloudIdentifier[] CloudAssetIdentifiers { get; }
+	}
+
+	[Mac (10,13, onlyOn64 : true)]
+	[NoiOS][NoTV]
+	[Protocol]
+	interface PHContentEditingController {
+
+		[Abstract]
+		[Export ("canHandleAdjustmentData:")]
+		bool CanHandleAdjustmentData (PHAdjustmentData adjustmentData);
+
+		[Abstract]
+		[Export ("startContentEditingWithInput:placeholderImage:")]
+		void StartContentEditing (PHContentEditingInput contentEditingInput, NSImage placeholderImage);
+
+		[Abstract]
+		[Export ("finishContentEditingWithCompletionHandler:")]
+		void FinishContentEditing (Action<PHContentEditingOutput> completionHandler);
+
+		[Abstract]
+		[Export ("cancelContentEditing")]
+		void CancelContentEditing ();
+
+		[Abstract]
+		[Export ("shouldShowCancelConfirmation")]
+		bool ShouldShowCancelConfirmation { get; }
+	}
+#endif
 }
